strong verification on memory allocation efficiency and leakage

+++ ymir1.cc
@@ -123,9 +123,9 @@ static unsigned int
 ymir_option_lang_mask (void) {
     return CL_YMIR;
 }
-
+    
 static bool
-ymir_langhook_handle_option (size_t scode, const char *arg, int value ATTRIBUTE_UNUSED,
+ymir_langhook_handle_option (size_t scode, const char *arg, long int value ATTRIBUTE_UNUSED,
 		   int kind ATTRIBUTE_UNUSED, location_t loc ATTRIBUTE_UNUSED,
 		   const struct cl_option_handlers *handlers ATTRIBUTE_UNUSED)
 {
